refactor: Moved validation code
The validation of the notary files were moved to `validate.py`, so `notary_api.py` only consists of API related functions.
