+++ connaisseur/tests/test_notary_api.py
@@ -9,55 +9,7 @@ import connaisseur.trust_data
 import connaisseur.notary_api as notary_api
 from connaisseur.image import Image
 from connaisseur.tuf_role import TUFRole
-from connaisseur.key_store import KeyStore
 from connaisseur.exceptions import BaseConnaisseurException
-
-policy_rule1 = {
-    "pattern": "docker.io/securesystemsengineering/alice-image",
-    "verify": True,
-    "delegations": ["phbelitz", "chamsen"],
-}
-policy_rule2 = {"pattern": "docker.io/securesystemsengineering/*:*", "verify": True}
-targets1 = {
-    "v1": {
-        "hashes": {"sha256": "E4irx6ElMoNsOoG9sAh0CbFSCPWuunqHrtz9VtY3wUU="},
-        "length": 1994,
-    },
-    "v2": {
-        "hashes": {"sha256": "uKOFIodqniVQ1YLOUaHYfr3GxXDl5YXQhWC/1kb3+AQ="},
-        "length": 1994,
-    },
-}
-
-req_delegations1 = ["targets/phbelitz", "targets/chamsen"]
-req_delegations2 = []
-req_delegations3 = ["targets/someuserthatdidnotsign"]
-targets1 = [
-    {
-        "test": {
-            "hashes": {"sha256": "rJBMmxkdFPr1S3lS8mUKS7IcIBvzQTE4i4UejOmSplI="},
-            "length": 1993,
-        }
-    },
-    {
-        "test": {
-            "hashes": {"sha256": "rJBMmxkdFPr1S3lS8mUKS7IcIBvzQTE4i4UejOmSplI="},
-            "length": 1993,
-        }
-    },
-]
-targets2 = [
-    {
-        "sign": {
-            "hashes": {"sha256": "oVR5e4MAFllW7h8W2Y86FCYwHBFo8EYsc86bwDNhyr8="},
-            "length": 1994,
-        },
-        "v1": {
-            "hashes": {"sha256": "eZwPqKpMn7/1qZrvG0tcOrucLzQTQ0UAWYL600iYk8c="},
-            "length": 1994,
-        },
-    }
-]
 
 
 @pytest.fixture
@@ -152,20 +104,6 @@ def mock_request(monkeypatch):
 
 
 @pytest.fixture
-def mock_keystore(monkeypatch):
-    def init(self):
-        self.keys = {
-            "root": (
-                "MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEtR5kwrDK22SyCu7WMF8tCjVgeORA"
-                "S2PWacRcBN/VQdVK4PVk1w4pMWlz9AHQthDGl+W2k3elHkPbR+gNkK2PCA=="
-            )
-        }
-        self.hashes = {}
-
-    monkeypatch.setattr(KeyStore, "__init__", init)
-
-
-@pytest.fixture
 def mock_trust_data(monkeypatch):
     def _validate_expiry(self):
         pass
@@ -207,100 +145,6 @@ def test_health_check_acr(acrapi, mock_request, host: str, out: bool):
 def test_is_notary_selfsigned(napi, slfsig: str, out: bool):
     os.environ["SELFSIGNED_NOTARY"] = slfsig
     assert napi.is_notary_selfsigned() == out
-
-
-@pytest.mark.parametrize(
-    "image, policy_rule, digest",
-    [
-        (
-            "securesystemsengineering/alice-image:test",
-            policy_rule1,
-            "ac904c9b191d14faf54b7952f2650a4bb21c201bf34131388b851e8ce992a652",
-        ),
-        (
-            (
-                (
-                    "securesystemsengineering/alice-image@sha256:ac904c9b191d14faf54b7952f2650a4bb21"
-                    "c201bf34131388b851e8ce992a652"
-                )
-            ),
-            policy_rule1,
-            "ac904c9b191d14faf54b7952f2650a4bb21c201bf34131388b851e8ce992a652",
-        ),
-        (
-            "securesystemsengineering/sample-image:sign",
-            policy_rule2,
-            "a154797b8300165956ee1f16d98f3a1426301c1168f0462c73ce9bc03361cabf",
-        ),
-        (
-            "securesystemsengineering/sample-image:v1",
-            policy_rule2,
-            "799c0fa8aa4c9fbff5a99aef1b4b5c3abb9c2f34134345005982fad3489893c7",
-        ),
-    ],
-)
-def test_get_trusted_digest(
-    napi,
-    mock_trust_data,
-    mock_keystore,
-    mock_request,
-    image: str,
-    policy_rule: dict,
-    digest: str,
-):
-    assert napi.get_trusted_digest("host", Image(image), policy_rule) == digest
-
-
-def test_get_trusted_digest_error():
-    pass
-
-
-@pytest.mark.parametrize(
-    "image, req_delegations, targets",
-    [
-        ("securesystemsengineering/alice-image", req_delegations1, targets1),
-        ("securesystemsengineering/sample-image", req_delegations2, targets2),
-    ],
-)
-def test_process_chain_of_trust(
-    napi,
-    mock_keystore,
-    mock_request,
-    mock_trust_data,
-    image: str,
-    req_delegations: dict,
-    targets: list,
-):
-    assert napi.process_chain_of_trust("host", Image(image), req_delegations) == targets
-
-
-@pytest.mark.parametrize(
-    "image, req_delegations, error",
-    [
-        (
-            "docker.io/securesystemsengineering/sample-image",
-            req_delegations1,
-            "could not find any delegations in trust data.",
-        ),
-        (
-            "securesystemsengineering/alice-image",
-            req_delegations3,
-            "could not find delegation roles ['targets/someuserthatdidnotsign'] in trust data.",
-        ),
-    ],
-)
-def test_process_chain_of_trust_error(
-    napi,
-    mock_keystore,
-    mock_request,
-    mock_trust_data,
-    image: str,
-    req_delegations: list,
-    error: str,
-):
-    with pytest.raises(BaseConnaisseurException) as err:
-        napi.process_chain_of_trust("host", Image(image), req_delegations)
-    assert error in str(err.value)
 
 
 @pytest.mark.parametrize(
@@ -449,53 +293,3 @@ def test_get_auth_token_error_acr(acrapi, mock_request, url: str, error: str):
     with pytest.raises(BaseConnaisseurException) as err:
         acrapi.get_auth_token(url)
     assert error in str(err.value)
-
-
-@pytest.mark.parametrize(
-    "image, digest",
-    [
-        (
-            (
-                "image@sha256:1388abc7a12532836c3a81"
-                "bdb0087409b15208f5aeba7a87aedcfd56d637c145"
-            ),
-            "1388abc7a12532836c3a81bdb0087409b15208f5aeba7a87aedcfd56d637c145",
-        ),
-        (
-            (
-                "image@sha256:b8a38522876a9e2550d582"
-                "ce51a1d87ebdc6c570e5e585d08560bfd646f7f804"
-            ),
-            "b8a38522876a9e2550d582ce51a1d87ebdc6c570e5e585d08560bfd646f7f804",
-        ),
-        (
-            (
-                "image@sha256:b8a38522876a9e2550d582"
-                "ce51a1d87ebdc6c570e5e585d08560bfd646f7f805"
-            ),
-            None,
-        ),
-    ],
-)
-def test_search_image_targets_for_digest(napi, image: str, digest: str):
-    data = trust_data("tests/data/sample_releases.json")["signed"]["targets"]
-    assert napi.search_image_targets_for_digest(data, Image(image)) == digest
-
-
-@pytest.mark.parametrize(
-    "image, digest",
-    [
-        (
-            "image:v1",
-            "1388abc7a12532836c3a81bdb0087409b15208f5aeba7a87aedcfd56d637c145",
-        ),
-        (
-            "image:v2",
-            "b8a38522876a9e2550d582ce51a1d87ebdc6c570e5e585d08560bfd646f7f804",
-        ),
-        ("image:v3", None),
-    ],
-)
-def test_search_image_targets_for_tag(napi, image: str, digest: str):
-    data = trust_data("tests/data/sample_releases.json")["signed"]["targets"]
-    assert napi.search_image_targets_for_tag(data, Image(image)) == digest
